pyproject.toml
[project]
name = "jaxoracer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "av>=17.0.0",
    "jax>=0.9.2",
    "opencv-python>=4.13.0.92",
    "pyyaml>=6.0.3",
    "typer>=0.24.1",
    "wandb>=0.26.0",
]
